attempt to correct the reading angles from the MPU6050 sensor

+++ VulCAN-Proj/lib/arduino-CAN-master/src/ESP32_CAN_CTRL.h
@@ -5,7 +5,7 @@
 #include <CAN.h>
 
 void CAN_Setup(int rx, int tx);
-void sendCommands(int Z, int X);
+void sendCommands(float Z, float X);
 void CAN_Sender();
 void CAN_Receiver();
 void parseMessage(char* data);
